Move `getElementId` to `helpers`
fix
