+++ packages/components/addon/components/hds/flyout/index.js
@@ -6,7 +6,7 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { assert } from '@ember/debug';
-import { getElementId } from '../form/utils/getElementId';
+import { getElementId } from '@hashicorp/design-system-components/helpers/hds-get-element-id';
 import { tracked } from '@glimmer/tracking';
 
 export const DEFAULT_SIZE = 'medium';
