+++ packages/components/addon/components/hds/form/fieldset/index.js
@@ -6,8 +6,8 @@
 import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
-import { getElementId } from '../utils/getElementId';
-import { setAriaDescribedBy } from '../utils/setAriaDescribedBy';
+import { getElementId } from '@hashicorp/design-system-components/helpers/hds-get-element-id';
+import { setAriaDescribedBy } from '../../../../helpers/hds-set-aria-described-by';
 
 export default class HdsFormFieldsetIndexComponent extends Component {
   @tracked ariaDescribedBy = this.args.extraAriaDescribedBy;
